+++ src/pages/GettingStarted/CreateTenant-Section.vue
